pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-builder"
version = "1.0.0"
description = "Visual agent builder for TFrameX - Create sophisticated LLM agents and workflows"
authors = [
    { name = "TesslateAI", email = "contact@tesslate.com" }
]
dependencies = [
    # "tframex==1.1.0",  # Uncomment when 1.1.0 is published to PyPI
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.1",
    "aiohttp>=3.9.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/TesslateAI/Agent-Builder"
Documentation = "https://github.com/TesslateAI/Agent-Builder/wiki"
Repository = "https://github.com/TesslateAI/Agent-Builder"
Issues = "https://github.com/TesslateAI/Agent-Builder/issues"

[tool.setuptools.packages.find]
where = ["builder"]
include = ["backend*", "frontend*"]

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true